refactor: move resolver, mutator and subject-type into server module
